simplify solar adjustment
